Last fixes concerning 64 to 32 bits + reduce the voronoidzone test iteration + fix a test that was using a null range

+++ oscar-cbls/src/main/scala/oscar/cbls/business/routing/model/VRP.scala
@@ -241,9 +241,9 @@ class VRP(val m: Store, val n: Int, val v: Int, maxPivotPerValuePercent:Int = 4)
       toReturn
   }
 
-  def stringOfVehicle(vehicle:Long):Option[String] = {
+  def stringOfVehicle(vehicle:Int):Option[String] = {
     val routeOfV = getRouteOfVehicle(vehicle)
-    if(routeOfV.length == 1L){
+    if(routeOfV.length == 1){
       None
     }else{
       Some("" + routeOfV.mkString("->") + "->" + vehicle)
